Add a pytest
This test checks whether the maximum error in the oscillations of the laser width (due to mismatching) is below 4% in comparison to the theory. The variables z_per_pl and r_per_pw are chosen in such a way that the error is relatively small, and the calculation is relatively fast as well.